refactor(table): add/replace missing 'type' in imports
