+++ src/shared/components/Table/Body/Line/Line.tsx
@@ -2,7 +2,7 @@ import { Table as TableTk } from '@axa-fr/react-toolkit-all';
 import { ComponentPropsWithoutRef } from 'react';
 import Header from '../../Headers/Header';
 import { type Tcol } from '../types';
-import Td, { TTdContainer } from './Td';
+import Td, { type TTdContainer } from './Td';
 
 export type TLine = ComponentPropsWithoutRef<typeof TableTk.Tr> & {
   columns: (TTdContainer & Tcol & { keyCol: string })[];
