+++ src/shared/components/Table/Headers/Header/types.d.ts
@@ -1,4 +1,4 @@
-import { ReactNode } from 'react';
+import { type ReactNode } from 'react';
 
 export type THeader = {
   key: string;
