+++ src/shared/components/Table/index.ts
@@ -3,8 +3,7 @@ export { default as Line } from './Body/Line';
 export { default as Td } from './Body/Line/Td';
 export { default as Headers } from './Headers';
 export { default as Header } from './Headers/Header';
-export { default as Th } from './Headers/Header/Th';
-export type { Torder } from './Headers/Header/Th';
+export { default as Th, type Torder } from './Headers/Header/Th';
 export { default } from './Table.container';
 export { ASCENDING, DESCENDING, NONE } from './constants';
 export { setDisplay, setTextTooltip } from './formatDataTable';
